move classes into configuration
